Documents/TelegramBot/requirements.txt
python-telegram-bot==20.3
telethon==1.27.0
nest_asyncio==1.5.6